since SDK4 kept getting the following error:
Exception: Glyph too large! codepoint 87: 276 > 256. Font /Users/obakesan/CloudStation/Projects/Pebble/OnePlus Life/resources/fonts/Rokkitt-Bold.ttf

as a fix use the regex for font resrouce Rokkitt to only extract the numbers 0-9 which are used for 1 in the logo and the day. This seems to fix the issue by reducing the character set and it reduces the memory footprint which is better anyway.